OCR/docqa/pyproject.toml
[tool.poetry]
package-mode = false
name = "docqa"
version = "0.1.0"
description = "Digitize scanned PDFs"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
python-dotenv = "^1.0.1"
#transformers = {extras = ["torch"], version = "^4.39.3"}
pillow = "^10.3.0"
pypdf = "^3.0.1"
pymupdf = "^1.24.2"
pdf2image = "^1.17.0"
pytest-mock = "^3.14.0"
torchvision = "^0.19.0"
scipy = "^1.13.0"
scikit-learn = "^1.4.2"
levenshtein = "^0.25.1"
lxml = "^5.2.1"
docopt = "^0.6.2"
cdifflib = "^1.2.6"
fastapi = {extras = ["standard"], version = "^0.112.1"}
python-fasthtml = "^0.6.8"
sentencepiece = "^0.2.0"
protobuf = "^5.28.2"
httpx = "^0.27.0"
donut-python = "^1.0.9"
transformers = "^4.38.2"
torch = "^2.4.0"
pytorch-lightning = "^1.6.4"
timm = "^0.5.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pytest = "^8.1.1"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "docq.main:main"

[tool.ruff]
line-length = 118
target-version = "py310"
